+++ changelog.md
@@ -4,6 +4,10 @@
 
 - Added the `cai_causal_graph.identify_utils.identify_confounders` utility function, which allows you to identify
   a list of confounders between two nodes in a `cai_causal_graph.causal_graph.CausalGraph`.
+- Added the `cai_causal_graph.identify_utils.identify_instruments` utility function, which allows you to identify
+  a list of instrumental variables between two nodes in a `cai_causal_graph.causal_graph.CausalGraph`.
+- Added the `cai_causal_graph.identify_utils.identify_mediators` utility function, which allows you to identify
+  a list of mediators between two nodes in a `cai_causal_graph.causal_graph.CausalGraph`.
 
 ## 0.2.4